Rename the AES Codec package to aes
